+++ csentence.h
@@ -43,7 +43,7 @@ namespace ukb {
 	size_type size() const {return m_syns.size(); }
 
 	const std::string & syn(size_t i) const { return m_syns[i];}
-	float rank(size_t i) const { return m_ranks[i];}
+	double rank(size_t i) const { return m_ranks[i];}
 
 	std::string word() const { return w; }
 
@@ -61,7 +61,7 @@ namespace ukb {
 
 	void empty_synsets() { 
 	  std::vector<std::string>().swap(m_syns); 
-	  std::vector<float>().swap(m_ranks); 
+	  std::vector<double>().swap(m_ranks); 
 	  m_disamb = false;
 	}
 	std::vector<std::string> & get_syns_vector() { return m_syns; }
@@ -95,7 +95,7 @@ namespace ukb {
 	char m_pos; // 'n', 'v', 'a', 'r' or 0 (no pos)
 	float m_weight;     // Initial weight for PPV
 	std::vector<std::string> m_syns;
-	std::vector<float> m_ranks;
+	std::vector<double> m_ranks;
 	bool m_is_synset; // Wether the cword is just a synset (created with create_synset_cword)
 	bool m_distinguished;
 	bool m_disamb;      // If word is disambiguated, that is, if the synset
@@ -152,15 +152,15 @@ namespace ukb {
   };
 
   bool calculate_kb_ppr(const CSentence & cs,
-						std::vector<float> & res,
+						std::vector<double> & res,
 						bool with_weight);
   
   void calculate_kb_ppr_by_word_and_disamb(CSentence & cs,
 										   bool with_weight);
 
-  bool calculate_kb_ppv_csentence(CSentence & cs, std::vector<float> & res);
+  bool calculate_kb_ppv_csentence(CSentence & cs, std::vector<double> & res);
 
   void disamb_csentence_kb(CSentence & cs,
-							std::vector<float> & ranks);
+							std::vector<double> & ranks);
 }
 #endif
